requirements.txt
# Bangalore Transit Data Pipeline Dependencies

# Kafka Integration
kafka-python==2.0.2

# Pathway Engine for real-time data processing
pathway==0.7.6

# Data Processing and Analysis
pandas==2.1.4
numpy==1.24.3

# JSON and Configuration
pydantic==2.5.2
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# Date and Time Handling
python-dateutil==2.8.2

# HTTP Requests (for future API integrations)
requests==2.31.0

# Web Server
Flask==3.0.0
Flask-CORS==4.0.0

# Visualization (optional)
matplotlib==3.8.2
plotly==5.17.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0